Move PluginConfig into config package
