+++ src/main/kotlin/com/projectcitybuild/core/infrastructure/database/DataSourceProvider.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.core.infrastructure.database
 
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.modules.config.PlatformConfig
 import com.projectcitybuild.modules.logger.PlatformLogger
 import dagger.Module
