+++ src/main/kotlin/com/projectcitybuild/features/teleporthistory/listeners/PlayerSummonListener.kt
@@ -2,7 +2,7 @@ package com.projectcitybuild.features.teleporthistory.listeners
 
 import com.projectcitybuild.core.SpigotListener
 import com.projectcitybuild.entities.CrossServerLocation
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.features.teleporthistory.repositories.LastKnownLocationRepositoy
 import com.projectcitybuild.features.teleporting.events.PlayerPreSummonEvent
 import com.projectcitybuild.modules.config.PlatformConfig
