+++ src/main/kotlin/com/projectcitybuild/features/warps/commands/WarpCommand.kt
@@ -3,7 +3,7 @@ package com.projectcitybuild.features.warps.commands
 import com.projectcitybuild.core.InvalidCommandArgumentsException
 import com.projectcitybuild.core.utilities.Failure
 import com.projectcitybuild.core.utilities.Success
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.features.warps.repositories.WarpRepository
 import com.projectcitybuild.features.warps.usecases.WarpUseCase
 import com.projectcitybuild.modules.config.PlatformConfig
