+++ src/main/kotlin/com/projectcitybuild/features/warps/usecases/WarpListUseCase.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.features.warps.usecases.warplist
 
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.features.warps.repositories.WarpRepository
 import com.projectcitybuild.modules.config.PlatformConfig
 import javax.inject.Inject
