+++ src/main/kotlin/com/projectcitybuild/modules/config/PluginConfig.kt
@@ -1,4 +1,4 @@
-package com.projectcitybuild.entities
+package com.projectcitybuild.modules.config
 
 sealed class PluginConfig {
