+++ src/main/kotlin/com/projectcitybuild/modules/datetime/DateTimeProvider.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.modules.datetime
 
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.modules.config.PlatformConfig
 import com.projectcitybuild.modules.datetime.formatter.DateTimeFormatter
 import com.projectcitybuild.modules.datetime.formatter.DateTimeFormatterImpl
