+++ src/main/kotlin/com/projectcitybuild/modules/sharedcache/adapters/FlatFileSharedCacheSet.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.modules.sharedcache.adapters
 
-import com.projectcitybuild.entities.PluginConfig
+import com.projectcitybuild.modules.config.PluginConfig
 import com.projectcitybuild.modules.config.PlatformConfig
 import com.projectcitybuild.modules.sharedcache.SharedCacheSet
 import kotlinx.serialization.Serializable
